Create AddTagCommandTest.java
